refactor: ArrowRight 파일명 변경
