+++ frontend/src/components/@common/Icons/ArrowRightAlt.tsx
